requirements.txt
python-dotenv==1.0.0
pymongo==4.6.1
pyrogram==2.0.106
motor==3.4.0
tgcrypto==1.2.5
psutil
pyahocorasick